requirements.txt
streamlit
pandas
dtale
pycaret.regression
sklearn.preprocessing
matplotlib.pyplo
seaborn
